cmake instead of make, autogen instead of configure
SVN revision: 7501

+++ legacy/edje/conf-e-x11.sh
@@ -1,6 +1,9 @@
 #!/bin/sh
-
-cmake clean distclean
-rm config.cache
-./configure \
+if [ -f ./Makefile ]; then
+    make clean distclean
+fi
+if [ -f ./config.cache ]; then
+    rm config.cache
+fi
+./autogen.sh \
 $@
